services/newsletter/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain==0.1.16
langchain-openai==0.1.6
langchain-community==0.0.32
qdrant-client==1.7.0
python-dotenv==1.0.0
pydantic==2.5.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
tavily-python==0.3.1
google-search-results==2.4.2
requests==2.31.0
beautifulsoup4==4.12.2